Moved test to another file
