+++ lib/src/floodFilling.js
@@ -1,5 +1,6 @@
 /* eslint-disable */
 import * as THREE from 'three'
+import {bbAll} from './testdata/testSimpleBB.js'
 
 /*
 import {OBJLoader} from 'three/examples/jsm/loaders/OBJLoader'
@@ -10,18 +11,8 @@ loader.load(
   function ( object ) {
   },
 );
+
 */
-
-
-const box1 = new THREE.Box3(new THREE.Vector3(0, 0, 0), new THREE.Vector3(20, .3, 5))
-const box2 = new THREE.Box3(new THREE.Vector3(0, 0, 0), new THREE.Vector3(.3, 15, 5))
-const box3 = new THREE.Box3(new THREE.Vector3(9.7, 0, 0), new THREE.Vector3(10, 10, 5))
-const box4 = new THREE.Box3(new THREE.Vector3(0, 9.7, 0), new THREE.Vector3(15, 10, 5))
-const box5 = new THREE.Box3(new THREE.Vector3(0, 5, 0), new THREE.Vector3(5, 5.3, 5))
-const box6 = new THREE.Box3(new THREE.Vector3(5, 0, 0), new THREE.Vector3(5, 5.3, 5))
-const box7 = new THREE.Box3(new THREE.Vector3(19.7, 0, 0), new THREE.Vector3(20, 15, 5))
-const box8 = new THREE.Box3(new THREE.Vector3(0, 14.7, 0), new THREE.Vector3(20, 15, 5))
-const box9 = new THREE.Box3(new THREE.Vector3(15, 9.7, 0), new THREE.Vector3(15.3, 15, 5))
 
 
 export let pointsAccepted = []
@@ -41,10 +32,9 @@ const FFmaxAttempt = 20
 const voxelSize = 1
 const maxThreshold = 10000
 
-
-export const bbAll = [box1, box2, box3, box4, box5, box6, box7, box8, box9]
-
-RecursiveFloodFilling(bbAll)
+// UNCOMMENT FOR HTML TEST RENDERING
+//
+//RecursiveFloodFilling(bbAll)
 
 export async function FloodFilling(boundingBoxesAll) {
   const IFCUnionBox = calculateIFCBoundingBox(boundingBoxesAll)
